Revert "Class 10 Websockets"
