+++ package.json
@@ -20,8 +20,6 @@
   "homepage": "https://github.com/IsraQuirozZ/server-express#readme",
   "dependencies": {
     "express": "^4.18.2",
-    "express-handlebars": "^7.0.7",
-    "multer": "^1.4.5-lts.1",
-    "socket.io": "^4.6.1"
+    "express-handlebars": "^7.0.7"
   }
 }
