--- public/form.html
@@ -1,113 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link
-      href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha3/dist/css/bootstrap.min.css"
-      rel="stylesheet"
-      integrity="sha384-KK94CHFLLe+nY2dmCWGMq91rCGa5gtU4mk92HdvYe+M/SXH301p5ILy+dN9+nJOZ"
-      crossorigin="anonymous"
-    />
-    <link rel="stylesheet" href="/public/styles.css" />
-    <title>FORM</title>
-  </head>
-  <body>
-    <nav class="navbar navbar-expand-sm bg-danger">
-      <div class="container-fluid">
-        <a class="navbar-brand text-light" href="#">Home</a>
-        <button
-          class="navbar-toggler"
-          type="button"
-          data-bs-toggle="collapse"
-          data-bs-target="#navbarNav"
-          aria-controls="navbarNav"
-          aria-expanded="false"
-          aria-label="Toggle navigation"
-        >
-          <span class="navbar-toggler-icon"></span>
-        </button>
-        <div class="collapse navbar-collapse" id="navbarNav">
-          <ul class="navbar-nav">
-            <li class="nav-item">
-              <a class="nav-link text-light" href="#">Products</a>
-            </li>
-            <li class="nav-item">
-              <a class="nav-link text-light" href="#">Chat</a>
-            </li>
-          </ul>
-        </div>
-      </div>
-    </nav>
-    <main class="d-flex justify-content-center align-items-center">
-      <div class="container">
-        <div class="row">
-          <div class="col-6 d-flex flex-column">
-            <h3>Product register</h3>
-            <form id="productsForm" action="/new-product" , method="POST">
-              <fieldset class="mb-2">
-                <label for="productTitle" class="form-label">Title</label>
-                <input
-                  type="text"
-                  class="form-control"
-                  name="title"
-                  id="productTitle"
-                />
-              </fieldset>
-              <fieldset class="mb-2">
-                <label for="productDescription" class="form-label"
-                  >Description</label
-                >
-                <input
-                  type="text"
-                  class="form-control"
-                  name="description"
-                  id="productDescription"
-                />
-              </fieldset>
-              <fieldset class="mb-2">
-                <label for="productPrice" class="form-label">Price</label>
-                <input
-                  type="number"
-                  class="form-control"
-                  name="price"
-                  id="productPrice"
-                />
-              </fieldset>
-              <fieldset class="mb-2">
-                <label for="productStock" class="form-label">Stock</label>
-                <input
-                  type="number"
-                  class="form-control"
-                  name="stock"
-                  id="productStock"
-                />
-              </fieldset>
-              <div class="mb-2">
-                <label for="productPhoto" class="form-label">Photo</label>
-                <input
-                  type="text"
-                  class="form-control"
-                  name="url_photo"
-                  id="productPhoto"
-                  aria-describedby="photoSource"
-                />
-              </div>
-              <button id="newProduct" type="submit" class="btn btn-danger mt-2">
-                Save product
-              </button>
-            </form>
-          </div>
-          <div id="productTable" class="col-6"></div>
-        </div>
-      </div>
-    </main>
-    <footer class="bg-warning p-2 text-center">E-COMMERCE</footer>
-  </body>
-  <script
-    src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha3/dist/js/bootstrap.bundle.min.js"
-    integrity="sha384-ENjdO4Dr2bkBIFxQpeoTz1HIcje39Wm4jDKdf19U8gI4ddQ3GYNS7NTKfAdVQSZe"
-    crossorigin="anonymous"
-  ></script>
-</html>
