+++ public/styles.css
@@ -1,22 +1,3 @@
-body {
-  width: 100%;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  align-items: center;
-}
-
-nav,
-main,
-footer {
-  width: 100%;
-}
-main {
-  flex-grow: 1;
-  background-color: rgb(255, 160, 160);
-}
-
 img {
-  object-fit: contain;
+  object-fit: cover;
 }
