--- src/data/carts.json
@@ -1,11 +1,0 @@
-[
-  {
-    "products": [
-      {
-        "id": 2,
-        "stock": 10
-      }
-    ],
-    "id": 2
-  }
-]
