--- src/routes/api/products.js
@@ -1,68 +1,0 @@
-import { Router } from "express";
-import manager from "./../../managers/Product.js";
-
-const router = Router();
-
-router.post("/", async (req, res, next) => {
-  try {
-    let response = await manager.add_product(req.body);
-    if (response === 201) {
-      return res.json({ status: 201, message: "product created" });
-    }
-    return res.json({ status: 400, message: "not created" });
-  } catch (error) {
-    next(error);
-  }
-});
-router.get("/", async (req, res, next) => {
-  try {
-    let products = manager.read_products();
-    if (products.length > 0) {
-      return res.json({ status: 200, products });
-    }
-    let message = "not found";
-    return res.json({ status: 404, message });
-  } catch (error) {
-    next(error);
-  }
-});
-router.get("/:pid", async (req, res, next) => {
-  try {
-    let id = Number(req.params.pid);
-    let product = manager.read_product(id);
-    if (product) {
-      return res.json({ status: 200, product });
-    }
-    let message = "not found";
-    return res.json({ status: 404, message });
-  } catch (error) {
-    next(error);
-  }
-});
-router.put("/:pid", async (req, res, next) => {
-  try {
-    let id = Number(req.params.pid);
-    let data = req.body;
-    let response = await manager.update_product(id, data);
-    if (response === 200) {
-      return res.json({ status: 200, message: "product updated" });
-    }
-    return res.json({ status: 404, message: "not found" });
-  } catch (error) {
-    next(error);
-  }
-});
-router.delete("/:pid", async (req, res, next) => {
-  try {
-    let id = Number(req.params.pid);
-    let response = await manager.destroy_product(id);
-    if (response === 200) {
-      return res.json({ status: 200, message: "product deleted" });
-    }
-    return res.json({ status: 404, message: "not found" });
-  } catch (error) {
-    next(error);
-  }
-});
-
-export default router;
