+++ src/views/index.handlebars
@@ -1,21 +1,22 @@
-<h3>HOLA {{name}}</h3>
-{{#if last_name}}
-  <h2>{{last_name}}</h2>
+{{#if lastName}}
+  <h2>Hola {{name}} {{lastName}}</h2>
 {{else}}
-  <h2>aca deberia haber un apellido, si existe</h2>
+  <h2>Hola {{name}}</h2>
 {{/if}}
+
 <div class="d-flex justify-content-evenly">
   {{#each alumnos}}
-    <div class="card" style="width: 15rem">
+    <div class="card" style="width: 18rem;">
       <img
         src={{this.photo}}
         class="card-img-top"
-        style="width: 15rem;height:8rem"
+        style="height: 8rem;"
+        alt="..."
       />
       <div class="card-body">
         <h5 class="card-title">{{this.name}}</h5>
-        <p class="card-text">Alumno de backend.</p>
-        <button class="emit_data" class="btn btn-primary">+Info</button>
+        <p class="card-text">Alumno de backend</p>
+        <a href="#" class="btn btn-primary">+ info</a>
       </div>
     </div>
   {{/each}}
